Fix Vercel routing for SPA
This is a fix for 404 error when a page aside / is refreshed.
This error happens because vercel tends to find a page with
the exact path which doesn't exist because the paths were created
by dynamic routing in react router. The solution is to tell vercel
to give us only the root '/' and leave the rest to react router.